Cast this value since a uint64_t is not the same as an unsigned long long on a 64-bit machine.
Reported by kpfleming on IRC



git-svn-id: https://origsvn.digium.com/svn/asterisk/trunk@157874 65c4cc65-6c06-0410-ace0-fbb531ad65f3

+++ res/res_timing_timerfd.c
@@ -158,7 +158,7 @@ static void timerfd_timer_ack(int handle, unsigned int quantity)
 	} while (read_result != sizeof(expirations));
 
 	if (expirations != quantity) {
-		ast_debug(2, "Expected to acknowledge %u ticks but got %llu instead\n", quantity, expirations);
+		ast_debug(2, "Expected to acknowledge %u ticks but got %llu instead\n", quantity, (unsigned long long) expirations);
 	}
 }
